refactor: Split CAST tests from literals.rs into dedicated cast.rs module
Extract CAST function tests (lines 325-522) from literals.rs into a new
dedicated cast.rs module for better test organization and maintainability.

Changes:
- Created crates/parser/src/tests/cast.rs with 13 CAST function tests
- Updated crates/parser/src/tests/mod.rs to include cast module
- Removed CAST tests from literals.rs (reduced from 522 to 323 lines)
- All 439 parser tests pass

Benefits:
- Clear separation: date/time literals vs type conversion functions
- Easier navigation: CAST-specific tests in dedicated file
- Aligns with SQL standards: CAST is a function, not a literal
