+++ crates/parser/src/tests/mod.rs
@@ -3,6 +3,7 @@ use super::*;
 mod aggregates;
 mod alter_table;
 mod case_expression;
+mod cast;
 mod create_table;
 mod cte;
 mod delete;
